[mtubio] Pantalla Ficha Cursos -- Tarjeta formadores implementada la funcionalidad tutor

+++ siga-core/src/main/java/org/itcgae/siga/DTOs/form/FormadorCursoItem.java
@@ -19,7 +19,7 @@ public class FormadorCursoItem {
 	private String nombre;
 	private String apellidos;
 	private Double tarifa;
-	private Double tutor;
+	private Short tutor;
 	private String nif;
 	private String tipoIdentificacion;
 	private Date fechaBaja;
@@ -233,17 +233,17 @@ public class FormadorCursoItem {
 	/**
 	 * 
 	 **/
-	public FormadorCursoItem tutor(Double tutor) {
+	public FormadorCursoItem tutor(Short tutor) {
 		this.tutor = tutor;
 		return this;
 	}
 
 	@JsonProperty("tutor")
-	public Double getTutor() {
+	public Short getTutor() {
 		return tutor;
 	}
 
-	public void setTutor(Double tutor) {
+	public void setTutor(Short tutor) {
 		this.tutor = tutor;
 	}
